refactor: move `getFilesRefList` to its own file
